Move animation classes to separate files
